requirements.txt
cryptography>=41.0.0
argon2-cffi>=23.1.0
pyperclip>=1.8.2
click>=8.1.0
requests>=2.31.0
pynacl>=1.5.0
python-dotenv>=1.0.0
colorama>=0.4.6